Extract and reuse chainable
