+++ addon-test-support/-private/execution_context/native-events-context.js
@@ -34,10 +34,6 @@ ExecutionContext.prototype = {
     throw new Error('not implemented');
   },
 
-  chainable() {
-    return this.pageObjectNode;
-  },
-
   click(selector, container) {
     const el = this.$(selector, container)[0];
     click(el);
